+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/forest/pruner/Pruner.java
@@ -28,6 +28,7 @@ import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -49,7 +50,8 @@ public class Pruner {
   private final AtomicReference<State> state = new AtomicReference<>(State.IDLE);
   private final ExecutorService executorService;
 
-  public Pruner(
+  @VisibleForTesting
+  Pruner(
       final MarkSweepPruner pruningStrategy,
       final Blockchain blockchain,
       final PrunerConfiguration prunerConfiguration,
